Create a plot datamation

Create a plot datamation from a tidyverse pipeline.

datamation_sanddance(
  pipeline,
  envir = rlang::global_env(),
  pretty = TRUE,
  elementId = NULL,
  height = 300,
  width = 300
)

Arguments

pipeline

Tidyverse pipeline

envir

Environment where code is evaluated. Defaults to the global environment.

pretty

Whether to pretty the JSON output. Defaults to TRUE.

elementId

Optional ID for the widget element.

height

Height of the plotting area of the widget (excluding axes and legends). This is an approximation and not an exact science, since sizes may vary depending on labels, legends, facets, etc! Defaults to 300 (pixels).

width

Width of the plotting area of the widget (excluding axes and legends). This is an approximation and not an exact science, since sizes may vary depending on labels, legends, facets, etc! Defaults to 300 (pixels).

Examples

{
  library(dplyr)

  "small_salary %>%
  group_by(Degree) %>%
  summarize(mean = mean(Salary))" %>%
    datamation_sanddance()

  library(ggplot2)

  "small_salary %>%
  group_by(Work, Degree) %>%
  summarize(mean_salary = mean(Salary)) %>%
  ggplot(aes(x = Work, y = mean_salary)) +
  geom_point() +
  facet_grid(rows = vars(Degree))" %>%
    datamation_sanddance()
}
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union